requirements.txt
py-clob-client==0.34.6
websockets
requests
python-dotenv
web3
